Share the demo UI components and base stylesheet.
